The hdf5 printer now calculates the highest point id on rank 0. The resume flag is now reset to false if there is no resume files.
	modified:   Printers/include/gambit/Printers/printers/hdf5printer.hpp
	modified:   Printers/src/printers/hdf5printer/hdf5printer.cpp
	modified:   ScannerBit/src/objectives/test_functions/examples.cpp
	modified:   ScannerBit/src/scanners/twalk/twalk.cpp
	modified:   Utils/include/gambit/Utils/mpiwrapper.hpp
	modified:   Utils/include/gambit/Utils/new_mpi_datatypes.hpp

+++ Utils/include/gambit/Utils/new_mpi_datatypes.hpp
@@ -81,13 +81,13 @@ namespace Gambit
     //typedef std::pair<unsigned long int, unsigned int> PPIDpair;
     struct PPIDpair
     {
-      long int pointID;
+      unsigned long long int pointID;
       unsigned int rank;
       PPIDpair() 
         : pointID(-1)
         , rank(0)
       {}
-      PPIDpair(const long int p, const int r)
+      PPIDpair(const unsigned long long int p, const int r)
         : pointID(p)
         , rank(r)
       {}
